#245: slight refactoring of repos
