+++ app/domain/dataapi/actions.py
@@ -1,5 +1,5 @@
 from app.data import model, repositories
-from app.data.repositories import layer2_repository
+from app.data.repositories import layer2
 from app.presentation import dataapi
 
 ENABLED_CATALOGS = [
@@ -17,20 +17,20 @@ class Actions(dataapi.Actions):
         filters = []
 
         if query.pgcs is not None:
-            filters.append(layer2_repository.PGCOneOfFilter(query.pgcs))
+            filters.append(layer2.PGCOneOfFilter(query.pgcs))
 
         if (query.ra is not None) and (query.dec is not None) and (query.radius is not None):
-            filters.append(layer2_repository.ICRSCoordinatesInRadiusFilter(query.ra, query.dec, query.radius))
+            filters.append(layer2.ICRSCoordinatesInRadiusFilter(query.ra, query.dec, query.radius))
 
         if query.name is not None:
-            filters.append(layer2_repository.DesignationCloseFilter(query.name, 3))
+            filters.append(layer2.DesignationCloseFilter(query.name, 3))
 
         if (query.cz is not None) and (query.cz_err_percent is not None):
-            filters.append(layer2_repository.RedshiftCloseFilter(query.cz, query.cz_err_percent))
+            filters.append(layer2.RedshiftCloseFilter(query.cz, query.cz_err_percent))
 
         objects = self.layer2_repo.query(
             ENABLED_CATALOGS,
-            layer2_repository.AndFilter(filters),
+            layer2.AndFilter(filters),
             query.page_size,
             query.page,
         )
